products/labmind/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sentence-transformers>=3.4.0
faiss-cpu>=1.9.0
numpy>=1.26.0
scipy>=1.14.0
pandas>=2.2.0
pydantic>=2.10.0
torch>=2.5.0
